MNT: Move METAR PEG parser to private subdirectory
This allows us to ignore it in CodeQL scanning.
