+++ .codeclimate.yml
@@ -40,8 +40,7 @@ plugins:
 
 exclude_patterns:
 - "src/metpy/io/_nexrad_msgs/msg*.py"
-- "src/metpy/io/metar_parse.peg"
-- "src/metpy/io/metar_parser.py"
+- "src/metpy/io/_metar_parser/*"
 - "docs/**/*"
 - "staticdata/**/*"
 - "tests/*/baseline/*"
