+++ setup.cfg
@@ -79,7 +79,7 @@ multiline-quotes = double
 rst-roles = class, data, doc, func, meth, mod
 rst-directives = plot, versionchanged
 docstring-convention = numpy
-exclude = docs build src/metpy/io/metar_parser.py
+exclude = docs build src/metpy/io/_metar_parser/metar_parser.py
 select = A B C D E F G H I J K L M N O P Q R S T U V W X Y Z B902
 ignore = F405 W503 RST902 SIM106
 per-file-ignores = examples/*.py: D MPY001 T003 T001
